closes #3: Add local variable to tangle config.org on save
Found a way to tangle files on save by adding a local variable eval,
to avoid the "unsafe local var" warning I added the exact same command to safe-list.
I also needed to add the property header for the tangle to actually fire, not sure
if needed?

I also updated the init.el file to load the config.el file if exists.